reimplement GiM idea about defining watches function...
WATCHER(x)		<- just normal watcher, we have got int as second param.
WATCHER_LINE(x)		<- it was WATCHER(x) but i don't really like typecasts like: (int) watch...
				2nd param (char *)
WATCHER_AUDIO(x) 	<- audio watches, we have got in 2nd param (string_t, name - buf)

you ought to use watch_add_line() to add line watches.. otherwise you will have gcc warning.

+++ ekg/scripts.c
@@ -721,7 +721,7 @@ WATCHER(script_handle_watch)
 	script_watch_t *temp = data;
 
 	SCRIPT_HANDLER_HEADER(script_handler_watch_t);
-	SCRIPT_HANDLER_FOOTER(script_handler_watch, type, fd, (int) watch) {
+	SCRIPT_HANDLER_FOOTER(script_handler_watch, type, fd, watch) {
 		if (!type) {
 			return -1; /* watch_free(temp->self); */
 		}
